Bunch of work for the map/chat area:
Displaying unique names in the chat area, when possible.
Showing a list of player's tokens in the bottom left, updated as users join
Support on the server and client to send a user's current position on the map to everyone else

Aside from some refactoring, and polish, this is the bare minimum needed to run a session.